requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
psycopg2
pytest
python-dotenv